# packages/schemas/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "SMR Twin Platform API",
    "version": "0.1.0",
    "description": "Simulation-only API for the SMR Twin Platform digital twin simulator. No real plant control."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local API Gateway"
    }
  ],
...
  "paths": {
...
    "/api/v1/system/status": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Read platform and integration status",
        "responses": {
          "200": {
            "description": "Current platform status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStatusResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/assets": {
      "get": {
        "tags": [
          "Assets"
        ],
        "summary": "List simulation assets",
        "responses": {
          "200": {
            "description": "Assets returned from simulation or labelled fallback memory source",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetsResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/telemetry/latest": {
      "get": {
        "tags": [
          "Telemetry"
        ],
        "summary": "Read latest telemetry points",
        "responses": {
          "200": {
            "description": "Latest telemetry points for SMR unit overview and thermal process loop",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetryLatestResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/telemetry/history": {
      "get": {
        "tags": [
          "Telemetry"
        ],
        "summary": "Read raw or downsampled telemetry history snapshots",
        "parameters": [
          {
            "name": "window",
            "in": "query",
            "required": false,
            "description": "History window supported by the simulation service.",
            "schema": {
              "type": "string",
              "enum": [
                "5m",
                "15m",
                "30m",
                "1h",
                "6h",
                "24h"
              ],
              "default": "15m"
            }
          },
          {
            "name": "resolution",
            "in": "query",
            "required": false,
            "description": "Telemetry history resolution. raw returns source samples; 1m returns synthetic historian downsampling buckets when available.",
            "schema": {
              "type": "string",
              "enum": [
                "raw",
                "1m"
              ],
              "default": "raw"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Raw or downsampled telemetry snapshots for synthetic simulation data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetryHistoryResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/commands": {
      "post": {
        "tags": [
          "Commands"
        ],
        "summary": "Submit a simulation-only command",
        "description": "Commands mutate only in-memory simulation state for V-101 and P-101. They do not control real equipment.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandRequest"
              },
              "examples": {
                "valveSetPosition": {
                  "value": {
                    "targetTag": "V-101",
                    "commandType": "SET_POSITION",
                    "payload": {
                      "positionPercent": 75
                    }
                  }
                },
                "pumpStart": {
                  "value": {
                    "targetTag": "P-101",
                    "commandType": "START",
                    "payload": {
                      "reason": "operator_demo"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command accepted, rejected, completed, or failed by the simulation service",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/commands/recent": {
      "get": {
        "tags": [
          "Commands"
        ],
        "summary": "List recent in-memory command records",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Maximum number of newest command records to return.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recent command records",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandsResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/events/recent": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "List recent unified in-memory events",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Maximum number of newest event records to return.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recent command, alarm, equipment, and simulation events",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/reports/simulation-summary": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Generate a simulation-only summary report",
        "description": "Exports synthetic simulator data for demo and portfolio use only. This is not a regulatory, production audit, or nuclear compliance report.",
        "parameters": [
          {
            "name": "window",
            "in": "query",
            "required": false,
            "description": "Report time window for telemetry statistics.",
            "schema": {
              "type": "string",
              "enum": [
                "15m",
                "1h",
                "6h",
                "24h"
              ],
              "default": "1h"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "description": "Response format. CSV returns text/csv and PDF returns application/pdf instead of the JSON envelope.",
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "csv",
                "pdf"
              ],
              "default": "json"
            }
          },
          {
            "name": "template",
            "in": "query",
            "required": false,
            "description": "Simulation-only presentation template. Templates change report layout/sections only and are not regulatory formats.",
            "schema": {
              "type": "string",
              "enum": [
                "executive-summary",
                "engineering-detail",
                "alarm-and-event-review",
                "pid-control-review",
                "historian-trend-summary"
              ],
              "default": "engineering-detail"
            }
          },
          {
            "name": "sections",
            "in": "query",
            "required": false,
            "description": "Comma-separated report sections to include. Safety boundary text remains mandatory in every format.",
            "schema": {
              "type": "string",
              "example": "metadata,systemSummary,pidSummary,trendStatistics"
            }
          },
          {
            "name": "includeDisclaimers",
            "in": "query",
            "required": false,
            "description": "Controls optional detailed disclaimer text only. Mandatory simulation-only/non-regulatory wording is never removed.",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Simulation-only report in JSON envelope or CSV export.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationReportResponse"
                }
              },
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              },
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/alarms/{alarmID}/acknowledge": {
      "post": {
        "tags": [
          "Alarms"
        ],
        "summary": "Acknowledge an active simulation alarm",
        "parameters": [
          {
            "name": "alarmID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcknowledgeAlarmRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Acknowledged alarm instance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlarmResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/simulation/scenarios": {
      "get": {
        "tags": [
          "Scenarios"
        ],
        "summary": "List predefined synthetic simulation scenarios",
        "responses": {
          "200": {
            "description": "Predefined simulation scenarios",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenariosResponse"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/simulation/scenarios/{scenarioName}/start": {
      "post": {
        "tags": [
          "Scenarios"
        ],
        "summary": "Start a predefined synthetic scenario",
        "parameters": [
          {
            "name": "scenarioName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated simulation status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationStatusResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "502": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/simulation/scenarios/stop": {
      "post": {
        "tags": [
          "Scenarios"
        ],
        "summary": "Stop the active synthetic scenario",
        "responses": {
          "200": {
            "description": "Updated simulation status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationStatusResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "502": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/simulation/reset": {
      "post": {
        "tags": [
          "Scenarios"
        ],
        "summary": "Reset the in-memory simulation",
        "responses": {
          "200": {
            "description": "Updated simulation status after reset",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationStatusResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "502": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/control/status": {
      "get": {
        "tags": [
          "Control"
        ],
        "summary": "Read simulation-only control mode and authority status",
        "description": "Returns the TIC-101 control mode used by command arbitration for V-101. This is simulation-only and does not control real equipment.",
        "responses": {
          "200": {
            "description": "Current control mode status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ControlStatusResponse"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/control/mode": {
      "post": {
        "tags": [
          "Control"
        ],
        "summary": "Set simulation-only control mode",
        "description": "Switches TIC-101 between MANUAL, AUTO, and DISABLED in memory. AUTO reserves V-101 for the simulation-only TIC-101 PID controller.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModeChangeRequest"
              },
              "examples": {
                "auto": {
                  "value": {
                    "mode": "AUTO",
                    "requestedBy": "demo-operator",
                    "reason": "Enable simulation-only PID demo"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated control status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ControlStatusResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/pid/status": {
      "get": {
        "tags": [
          "Control"
        ],
        "summary": "Get TIC-101 PID status",
        "responses": {
          "200": {
            "description": "Current simulation PID status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PIDStatusResponse"
                }
              }
            }
          },
          "503": {
            "description": "Simulation unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pid/config": {
      "patch": {
        "tags": [
          "Control"
        ],
        "summary": "Update TIC-101 PID configuration",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PIDConfigUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated simulation PID status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PIDStatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid PID configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "description": "Simulation unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/scenarios/validate": {
      "post": {
        "tags": [
          "Scenarios"
        ],
        "summary": "Validate a simulation-only YAML scenario draft",
        "description": "Validates YAML content for synthetic demo scenario authoring. The endpoint does not persist drafts, mutate the runtime registry, deploy scenarios, or control any real plant.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioValidationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Validation result for a simulation-only draft.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioValidationResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    }
  },
  "components": {
    "responses": {
      "Error": {
        "description": "Standard API error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "error",
          "meta"
        ],
        "additionalProperties": false
      },
...
      "SystemStatus": {
        "type": "object",
        "properties": {
          "platform": {
            "type": "string"
          },
          "mode": {
            "type": "string",
            "description": "Current platform mode. The MVP uses simulation_only."
          },
          "environment": {
            "type": "string"
          },
          "controlBoundary": {
            "type": "string",
            "description": "The MVP uses no_live_control."
          },
          "dataSource": {
            "type": "string",
            "description": "Honest source label such as synthetic_simulation or in_memory_fallback."
          },
          "backendApi": {
            "$ref": "#/components/schemas/ComponentStatus"
          },
          "mqttBroker": {
            "$ref": "#/components/schemas/ComponentStatus"
          },
          "simulationService": {
            "$ref": "#/components/schemas/ComponentStatus"
          },
          "historian": {
            "$ref": "#/components/schemas/ComponentStatus"
          },
          "simulationConnected": {
            "type": "boolean"
          },
          "simulationMode": {
            "type": "string"
          },
          "simulationHealth": {
            "type": "string"
          },
          "lastSimulationTimestamp": {
            "type": "string"
          },
          "safetyDisclaimer": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "platform",
          "mode",
          "environment",
          "controlBoundary",
          "dataSource",
          "backendApi",
          "mqttBroker",
          "simulationService",
          "historian",
          "simulationConnected",
          "safetyDisclaimer",
          "version",
          "timestamp"
        ],
        "additionalProperties": false
      },
...
      "CommandPayload": {
        "type": "object",
        "properties": {
          "positionPercent": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "reason": {
            "type": "string",
            "maxLength": 240
          }
        },
        "additionalProperties": false
      },
      "CommandRequest": {
        "type": "object",
        "properties": {
          "targetTag": {
            "type": "string",
            "enum": [
              "V-101",
              "P-101"
            ]
          },
          "commandType": {
            "type": "string",
            "enum": [
              "OPEN",
              "CLOSE",
              "STOP",
              "SET_POSITION",
              "START"
            ]
          },
          "source": {
            "type": "string"
          },
          "requestedBy": {
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/CommandPayload"
          },
          "correlationId": {
            "type": "string"
          }
        },
        "required": [
          "targetTag",
          "commandType"
        ],
        "additionalProperties": false
      },
...
      "AcknowledgeAlarmRequest": {
        "type": "object",
        "properties": {
          "acknowledgedBy": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "comment": {
            "type": "string",
            "maxLength": 500
          }
        },
        "additionalProperties": false
      },
...
      "SystemStatusResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SystemStatus"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "additionalProperties": false
      },
      "AssetsResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Asset"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "additionalProperties": false
      },
      "TelemetryLatestResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TelemetryPoint"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "additionalProperties": false
      },
      "TelemetryHistoryResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TelemetrySnapshot"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "additionalProperties": false
      },
      "CommandResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Command"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "additionalProperties": false
      },
      "CommandsResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Command"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "additionalProperties": false
      },
      "EventsResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Event"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "additionalProperties": false
      },
      "AlarmResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AlarmInstance"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "additionalProperties": false
      },
...
      "ScenariosResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Scenario"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "additionalProperties": false
      },
      "SimulationStatusResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SimulationStatus"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "additionalProperties": false
      },
...
      "ModeChangeRequest": {
        "type": "object",
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/ControlMode"
          },
          "requestedBy": {
            "type": "string",
            "maxLength": 120
          },
          "reason": {
            "type": "string",
            "maxLength": 300
          }
        },
        "required": [
          "mode"
        ],
        "additionalProperties": false
      },
...
      "ControlStatusResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ControlStatus"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "additionalProperties": false
      },
...
      "PIDConfigUpdateRequest": {
        "type": "object",
        "properties": {
          "setpoint": {
            "type": "number",
            "minimum": 270,
            "maximum": 310
          },
          "kp": {
            "type": "number",
            "minimum": 0
          },
          "ki": {
            "type": "number",
            "minimum": 0
          },
          "kd": {
            "type": "number",
            "minimum": 0
          },
          "outputMin": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "outputMax": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "requestedBy": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "PIDStatusResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PIDStatus"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "SimulationReportResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SimulationReport"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "additionalProperties": false
      },
...
      "ScenarioValidationRequest": {
        "type": "object",
        "properties": {
          "format": {
            "type": "string",
            "enum": [
              "yaml",
              "yml"
            ],
            "default": "yaml"
          },
          "content": {
            "type": "string",
            "maxLength": 65536
          }
        },
        "required": [
          "content"
        ],
        "additionalProperties": false
      },
...
      "ScenarioValidationResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ScenarioValidationResult"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "additionalProperties": false
      }
    }
  }
}